Latest Update
Another rewrite of CombatManager
Rewrote CastSpell()
Added MOTD support
Added support for Imms to add/remove languages to players
Added Cooking skill and trainer
Fixed several issues with buffs
Rewrote code to deal with potions

+++ src/Core/RoomManager.cs
@@ -321,7 +321,7 @@ namespace Kingdoms_of_Etrea.Core
             }
         }
 
-        internal bool AddGoldToRoom(uint rid, uint gp)
+        internal bool AddGoldToRoom(uint rid, ulong gp)
         {
             try
             {
@@ -338,7 +338,7 @@ namespace Kingdoms_of_Etrea.Core
             }
         }
 
-        internal bool GetGoldFromRoom(uint rid, uint gp)
+        internal bool GetGoldFromRoom(uint rid, ulong gp)
         {
             try
             {
